Make i18n feature contained in a duck
