+++ app/redux/constants/ActionTypes.js
@@ -3,7 +3,5 @@ import { generateConstants } from '../utils';
 export default generateConstants([
   'USERS_CLEAR_ERROR',
   'USERS_INDEX(ASYNC)',
-  'README_LOAD(ASYNC)',
-  'LOCALE_INITIALIZE',
-  'LOCALE_CHANGE(ASYNC)'
+  'README_LOAD(ASYNC)'
 ]);
