+++ app/redux/create.js
@@ -3,8 +3,11 @@ import { devTools, persistState } from 'redux-devtools';
 
 import createMiddleware from './clientMiddleware';
 import * as reducers from './reducers';
+import i18n from '../ducks/i18n';
 
-const reducer = combineReducers(reducers);
+// Once everythig is a duck we can create a app/duck/reducer.js file:
+// https://github.com/erikras/react-redux-universal-hot-example/blob/master/src/ducks/reducer.js
+const reducer = combineReducers({ i18n, ...reducers });
 
 export default function(client, data) {
   const middleware = createMiddleware(client);
