+++ app/redux/reducers/index.js
@@ -1,3 +1,2 @@
 export { default as users } from './users';
 export { default as readme } from './readme';
-export { default as i18n } from './i18n';
